Aligns sensor files with current coding style
