+++ app/src/main/java/org/fossasia/pslab/sensorfragment/SensorFragmentBH1750.java
@@ -1,4 +1,4 @@
-package org.fossasia.pslab.fragment;
+package org.fossasia.pslab.sensorfragment;
 
 import android.os.AsyncTask;
 import android.os.Bundle;
@@ -7,12 +7,13 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Spinner;
 import android.widget.TextView;
 
 import org.fossasia.pslab.R;
 import org.fossasia.pslab.communication.ScienceLab;
 import org.fossasia.pslab.communication.peripherals.I2C;
-import org.fossasia.pslab.communication.sensors.MLX90614;
+import org.fossasia.pslab.communication.sensors.BH1750;
 import org.fossasia.pslab.others.ScienceLabCommon;
 
 import java.io.IOException;
@@ -21,17 +22,16 @@ import java.io.IOException;
  * Created by asitava on 10/7/17.
  */
 
-public class SensorFragmentMLX90614 extends Fragment {
+public class SensorFragmentBH1750 extends Fragment {
     private ScienceLab scienceLab;
     private I2C i2c;
-    private SensorFragmentMLX90614.SensorDataFetch sensorDataFetch;
-    private TextView tvSensorMLX90614ObjectTemp;
-    private TextView tvSensorMLX90614AmbientTemp;
-    private MLX90614 MLX90614;
+    private SensorDataFetch sensorDataFetch;
+    private TextView tvSensorBH1750Luminosity;
+    private BH1750 sensorBH1750;
 
-    public static SensorFragmentMLX90614 newInstance() {
-        SensorFragmentMLX90614 sensorFragmentMLX90614 = new SensorFragmentMLX90614();
-        return sensorFragmentMLX90614;
+    public static SensorFragmentBH1750 newInstance() {
+        SensorFragmentBH1750 sensorFragmentBH1750 = new SensorFragmentBH1750();
+        return sensorFragmentBH1750;
     }
 
     @Override
@@ -39,15 +39,19 @@ public class SensorFragmentMLX90614 extends Fragment {
         super.onCreate(savedInstanceState);
         scienceLab = ScienceLabCommon.scienceLab;
         i2c = scienceLab.i2c;
-
+        try {
+            sensorBH1750 = new BH1750(i2c);
+        } catch (IOException | InterruptedException e) {
+            e.printStackTrace();
+        }
         Runnable runnable = new Runnable() {
             @Override
             public void run() {
                 while (true) {
                     if (scienceLab.isConnected()) {
                         try {
-                            sensorDataFetch = new SensorFragmentMLX90614.SensorDataFetch();
-                        } catch (IOException e) {
+                            sensorDataFetch = new SensorDataFetch();
+                        } catch (IOException | InterruptedException e) {
                             e.printStackTrace();
                         }
                         sensorDataFetch.execute();
@@ -60,29 +64,31 @@ public class SensorFragmentMLX90614 extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.sensor_mlx90614, container, false);
-
-        tvSensorMLX90614ObjectTemp = (TextView) view.findViewById(R.id.tv_sensor_mlx90614_object_temp);
-        tvSensorMLX90614AmbientTemp = (TextView) view.findViewById(R.id.tv_sensor_mlx90614_ambient_temp);
-
+        View view = inflater.inflate(R.layout.sensor_bh1750, container, false);
+        tvSensorBH1750Luminosity = (TextView) view.findViewById(R.id.tv_sensor_bh1750_luminosity);
+        Spinner spinnerSensorBH1750 = (Spinner) view.findViewById(R.id.spinner_sensor_bh1750);
+        try {
+            if (sensorBH1750 != null) {
+                sensorBH1750.setRange(spinnerSensorBH1750.getSelectedItem().toString());
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
         return view;
     }
 
 
     private class SensorDataFetch extends AsyncTask<Void, Void, Void> {
-        MLX90614 MLX90614 = new MLX90614(i2c);
-        Double dataMLX90614ObjectTemp;
-        Double dataMLX90614AmbientTemp;
+        Double dataBH1750;
 
-        private SensorDataFetch() throws IOException {
+        private SensorDataFetch() throws IOException, InterruptedException {
         }
 
         @Override
         protected Void doInBackground(Void... params) {
             try {
-                if (MLX90614 != null) {
-                    dataMLX90614ObjectTemp = MLX90614.getObjectTemperature();
-                    dataMLX90614AmbientTemp = MLX90614.getAmbientTemperature();
+                if (sensorBH1750 != null) {
+                    dataBH1750 = sensorBH1750.getRaw();
                 }
             } catch (IOException e) {
                 e.printStackTrace();
@@ -92,8 +98,7 @@ public class SensorFragmentMLX90614 extends Fragment {
 
         protected void onPostExecute(Void aVoid) {
             super.onPostExecute(aVoid);
-            tvSensorMLX90614ObjectTemp.setText(String.valueOf(dataMLX90614ObjectTemp));
-            tvSensorMLX90614AmbientTemp.setText(String.valueOf(dataMLX90614AmbientTemp));
+            tvSensorBH1750Luminosity.setText(String.valueOf(dataBH1750));
         }
     }
 }
