+++ app/src/main/java/org/fossasia/pslab/sensorfragment/SensorFragmentBMP180.java
@@ -1,4 +1,4 @@
-package org.fossasia.pslab.fragment;
+package org.fossasia.pslab.sensorfragment;
 
 import android.os.AsyncTask;
 import android.os.Bundle;
@@ -7,14 +7,12 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.EditText;
-import android.widget.Spinner;
 import android.widget.TextView;
 
 import org.fossasia.pslab.R;
 import org.fossasia.pslab.communication.ScienceLab;
 import org.fossasia.pslab.communication.peripherals.I2C;
-import org.fossasia.pslab.communication.sensors.TSL2561;
+import org.fossasia.pslab.communication.sensors.BMP180;
 import org.fossasia.pslab.others.ScienceLabCommon;
 
 import java.io.IOException;
@@ -23,20 +21,18 @@ import java.io.IOException;
  * Created by asitava on 10/7/17.
  */
 
-public class SensorFragmentTSL2561 extends Fragment {
+public class SensorFragmentBMP180 extends Fragment {
     private ScienceLab scienceLab;
     private I2C i2c;
     private SensorDataFetch sensorDataFetch;
-    private TextView tvSensorTSL2561FullSpectrum;
-    private TextView tvSensorTSL2561Infrared;
-    private TextView tvSensorTSL2561Visible;
-    private Spinner spinnerSensorTSL2561Gain;
-    private EditText etSensorTSL2561Timing;
-    private TSL2561 TSL2561;
+    private TextView tvSensorBMP180Temp;
+    private TextView tvSensorBMP180Altitude;
+    private TextView tvSensorBMP180Pressure;
+    private BMP180 sensorBMP180;
 
-    public static SensorFragmentTSL2561 newInstance() {
-        SensorFragmentTSL2561 sensorFragmentTSL2561 = new SensorFragmentTSL2561();
-        return sensorFragmentTSL2561;
+    public static SensorFragmentBMP180 newInstance() {
+        SensorFragmentBMP180 sensorFragmentBMP180 = new SensorFragmentBMP180();
+        return sensorFragmentBMP180;
     }
 
     @Override
@@ -45,7 +41,7 @@ public class SensorFragmentTSL2561 extends Fragment {
         scienceLab = ScienceLabCommon.scienceLab;
         i2c = scienceLab.i2c;
         try {
-            TSL2561 = new TSL2561(i2c);
+            sensorBMP180 = new BMP180(i2c);
         } catch (IOException | InterruptedException e) {
             e.printStackTrace();
         }
@@ -69,28 +65,18 @@ public class SensorFragmentTSL2561 extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.sensor_tsl2561, container, false);
-        tvSensorTSL2561FullSpectrum = (TextView) view.findViewById(R.id.tv_sensor_tsl2561_full);
-        tvSensorTSL2561Infrared = (TextView) view.findViewById(R.id.tv_sensor_tsl2561_infrared);
-        tvSensorTSL2561Visible = (TextView) view.findViewById(R.id.tv_sensor_tsl2561_visible);
-        spinnerSensorTSL2561Gain = (Spinner) view.findViewById(R.id.spinner_sensor_tsl2561_gain);
-        etSensorTSL2561Timing = (EditText) view.findViewById(R.id.et_sensor_tsl2561_timing);
+        View view = inflater.inflate(R.layout.sensor_bmp180, container, false);
 
-        try {
-            if (TSL2561 != null) {
-                TSL2561.setGain(spinnerSensorTSL2561Gain.getSelectedItem().toString());
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        tvSensorBMP180Temp = (TextView) view.findViewById(R.id.tv_sensor_bmp180_temp);
+        tvSensorBMP180Altitude = (TextView) view.findViewById(R.id.tv_sensor_bmp180_altitude);
+        tvSensorBMP180Pressure = (TextView) view.findViewById(R.id.tv_sensor_bmp180_pressure);
 
         return view;
     }
 
 
     private class SensorDataFetch extends AsyncTask<Void, Void, Void> {
-        TSL2561 TSL2561 = new TSL2561(i2c);
-        private int[] dataTSL2561;
+        double[] dataBMP180 = new double[3];
 
         private SensorDataFetch() throws IOException, InterruptedException {
         }
@@ -98,10 +84,10 @@ public class SensorFragmentTSL2561 extends Fragment {
         @Override
         protected Void doInBackground(Void... params) {
             try {
-                if (TSL2561 != null) {
-                    dataTSL2561 = TSL2561.getRaw();
+                if (sensorBMP180 != null) {
+                    dataBMP180 = sensorBMP180.getRaw();
                 }
-            } catch (IOException e) {
+            } catch (IOException | InterruptedException e) {
                 e.printStackTrace();
             }
             return null;
@@ -109,9 +95,9 @@ public class SensorFragmentTSL2561 extends Fragment {
 
         protected void onPostExecute(Void aVoid) {
             super.onPostExecute(aVoid);
-            tvSensorTSL2561FullSpectrum.setText(String.valueOf(dataTSL2561[0]));
-            tvSensorTSL2561Infrared.setText(String.valueOf(dataTSL2561[0]));
-            tvSensorTSL2561Visible.setText(String.valueOf(dataTSL2561[0]));
+            tvSensorBMP180Temp.setText(String.valueOf(dataBMP180[0]));
+            tvSensorBMP180Altitude.setText(String.valueOf(dataBMP180[1]));
+            tvSensorBMP180Pressure.setText(String.valueOf(dataBMP180[2]));
         }
     }
 }
