+++ app/src/main/java/org/fossasia/pslab/sensorfragment/SensorFragmentMain.java
@@ -1,4 +1,4 @@
-package org.fossasia.pslab.fragment;
+package org.fossasia.pslab.sensorfragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
