+++ app/src/main/java/org/fossasia/pslab/sensorfragment/SensorFragmentTSL2561.java
@@ -1,4 +1,4 @@
-package org.fossasia.pslab.fragment;
+package org.fossasia.pslab.sensorfragment;
 
 import android.os.AsyncTask;
 import android.os.Bundle;
@@ -7,13 +7,14 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.TextView;
 
 import org.fossasia.pslab.R;
 import org.fossasia.pslab.communication.ScienceLab;
 import org.fossasia.pslab.communication.peripherals.I2C;
-import org.fossasia.pslab.communication.sensors.BH1750;
+import org.fossasia.pslab.communication.sensors.TSL2561;
 import org.fossasia.pslab.others.ScienceLabCommon;
 
 import java.io.IOException;
@@ -22,16 +23,20 @@ import java.io.IOException;
  * Created by asitava on 10/7/17.
  */
 
-public class SensorFragmentBH1750 extends Fragment {
+public class SensorFragmentTSL2561 extends Fragment {
     private ScienceLab scienceLab;
     private I2C i2c;
     private SensorDataFetch sensorDataFetch;
-    private TextView tvSensorBH1750Luminosity;
-    private BH1750 BH1750;
+    private TextView tvSensorTSL2561FullSpectrum;
+    private TextView tvSensorTSL2561Infrared;
+    private TextView tvSensorTSL2561Visible;
+    private Spinner spinnerSensorTSL2561Gain;
+    private EditText etSensorTSL2561Timing;
+    private TSL2561 sensorTSL2561;
 
-    public static SensorFragmentBH1750 newInstance() {
-        SensorFragmentBH1750 sensorFragmentBH1750 = new SensorFragmentBH1750();
-        return sensorFragmentBH1750;
+    public static SensorFragmentTSL2561 newInstance() {
+        SensorFragmentTSL2561 sensorFragmentTSL2561 = new SensorFragmentTSL2561();
+        return sensorFragmentTSL2561;
     }
 
     @Override
@@ -40,7 +45,7 @@ public class SensorFragmentBH1750 extends Fragment {
         scienceLab = ScienceLabCommon.scienceLab;
         i2c = scienceLab.i2c;
         try {
-            BH1750 = new BH1750(i2c);
+            sensorTSL2561 = new TSL2561(i2c);
         } catch (IOException | InterruptedException e) {
             e.printStackTrace();
         }
@@ -64,23 +69,28 @@ public class SensorFragmentBH1750 extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.sensor_bh1750, container, false);
-        tvSensorBH1750Luminosity = (TextView) view.findViewById(R.id.tv_sensor_bh1750_luminosity);
-        Spinner spinnerSensorBH1750 = (Spinner) view.findViewById(R.id.spinner_sensor_bh1750);
+        View view = inflater.inflate(R.layout.sensor_tsl2561, container, false);
+        tvSensorTSL2561FullSpectrum = (TextView) view.findViewById(R.id.tv_sensor_tsl2561_full);
+        tvSensorTSL2561Infrared = (TextView) view.findViewById(R.id.tv_sensor_tsl2561_infrared);
+        tvSensorTSL2561Visible = (TextView) view.findViewById(R.id.tv_sensor_tsl2561_visible);
+        spinnerSensorTSL2561Gain = (Spinner) view.findViewById(R.id.spinner_sensor_tsl2561_gain);
+        etSensorTSL2561Timing = (EditText) view.findViewById(R.id.et_sensor_tsl2561_timing);
+
         try {
-            if (BH1750 != null) {
-                BH1750.setRange(spinnerSensorBH1750.getSelectedItem().toString());
+            if (sensorTSL2561 != null) {
+                sensorTSL2561.setGain(spinnerSensorTSL2561Gain.getSelectedItem().toString());
             }
         } catch (IOException e) {
             e.printStackTrace();
         }
+
         return view;
     }
 
 
     private class SensorDataFetch extends AsyncTask<Void, Void, Void> {
-        BH1750 BH1750 = new BH1750(i2c);
-        Double dataBH1750;
+        TSL2561 sensorTSL2561 = new TSL2561(i2c);
+        private int[] dataTSL2561;
 
         private SensorDataFetch() throws IOException, InterruptedException {
         }
@@ -88,8 +98,8 @@ public class SensorFragmentBH1750 extends Fragment {
         @Override
         protected Void doInBackground(Void... params) {
             try {
-                if (BH1750 != null) {
-                    dataBH1750 = BH1750.getRaw();
+                if (sensorTSL2561 != null) {
+                    dataTSL2561 = sensorTSL2561.getRaw();
                 }
             } catch (IOException e) {
                 e.printStackTrace();
@@ -99,7 +109,9 @@ public class SensorFragmentBH1750 extends Fragment {
 
         protected void onPostExecute(Void aVoid) {
             super.onPostExecute(aVoid);
-            tvSensorBH1750Luminosity.setText(String.valueOf(dataBH1750));
+            tvSensorTSL2561FullSpectrum.setText(String.valueOf(dataTSL2561[0]));
+            tvSensorTSL2561Infrared.setText(String.valueOf(dataTSL2561[0]));
+            tvSensorTSL2561Visible.setText(String.valueOf(dataTSL2561[0]));
         }
     }
 }
